+++ src/wam2layers/config.py
@@ -10,6 +10,7 @@ from pydantic import (
     ConfigDict,
     Field,
     FilePath,
+    HttpUrl,
     field_validator,
     model_validator,
 )
@@ -42,6 +43,9 @@ class BoundingBox(NamedTuple):
     south: Annotated[float, Field(ge=-80, le=80)]
     east: Annotated[float, Field(ge=-180, le=180)]
     north: Annotated[float, Field(ge=-80, le=80)]
+
+    def __str__(self):
+        return f"{self.west}, {self.south}, {self.east}, {self.north}"
 
 
 def validate_region(region):
@@ -89,7 +93,7 @@ class Config(BaseModel):
     for date 2022-07-15, variable u, and levtype "_ml" (note the underscore).
     """
 
-    preprocessed_data_folder: Path
+    preprocessed_data_folder: Union[HttpUrl, Path]
     """Location where the pre-processed data should be stored.
 
     If it does not exist, it will be created during pre-processing.
@@ -521,13 +525,15 @@ class Config(BaseModel):
     @classmethod
     def _expanduser(cls, path):
         """Resolve ~ in input paths."""
-        return path.absolute()
+        if isinstance(path, Path):
+            return path.absolute()
+        return path  # Leave URLs untouched
 
     @field_validator("preprocessed_data_folder", "output_folder")
     @classmethod
     def _make_dir(cls, path):
         """Create output dirs if they don't exist yet."""
-        if not path.exists():
+        if isinstance(path, Path) and not path.exists():
             logger.info(f"Creating output folder {path}")
             path.mkdir(parents=True)
         return path