+++ src/wam2layers/tracking/io.py
@@ -1,4 +1,7 @@
 import logging
+import time
+import traceback
+from copy import deepcopy
 from datetime import datetime as pydt
 from functools import lru_cache
 from pathlib import Path
@@ -9,7 +12,7 @@ import pandas as pd
 import xarray as xr
 
 from wam2layers import __version__
-from wam2layers.config import BoundingBox, Config
+from wam2layers.config import Config
 from wam2layers.preprocessing.utils import add_bounds
 from wam2layers.reference.variables import DATA_ATTRIBUTES
 from wam2layers.utils.calendar import CfDateTime, round_cftime
@@ -28,67 +31,106 @@ def output_path(date: CfDateTime, config: Config):
     return f"{output_dir}/{mode}_{date.strftime('%Y-%m-%dT%H-%M')}.nc"
 
 
-@lru_cache(maxsize=4)  # cache for speedup
-def read_data_at_date(
-    date: CfDateTime, input_dir: Path, tracking_domain: Optional[BoundingBox]
-):
-    """Load input data for given date.
+def fully_load_and_detach(ds: xr.Dataset) -> xr.Dataset:
+    """Force all variables into memory and detach from OPeNDAP/netCDF backend."""
+    # Extract data variables with explicit dimension names
+    loaded_vars = {var: (ds[var].dims, ds[var].values) for var in ds.data_vars}
 
-    This function is accessed every time step. By caching the result we reduce the
-    number of times the files need to be accessed.
-    """
-    file = input_path(date, input_dir)
-    ds = xr.open_dataset(file, cache=False, use_cftime=True)
-    if tracking_domain is not None:
-        return select_subdomain(ds, tracking_domain)
-    return ds
+    # Extract coords similarly (some coords are multi-dimensional)
+    loaded_coords = {
+        coord: (ds.coords[coord].dims, ds.coords[coord].values) for coord in ds.coords
+    }
+    return xr.Dataset(data_vars=loaded_vars, coords=loaded_coords, attrs=ds.attrs)
 
 
-def read_data_at_time(datetime: CfDateTime, config: Config) -> xr.Dataset:
-    """Get a single time slice from input data."""
-    ds = read_data_at_date(
-        datetime, config.preprocessed_data_folder, config.tracking_domain
-    )
-    return ds.sel(time=datetime, drop=True)
-
-
-def select_subdomain(ds: xr.Dataset, bbox: BoundingBox):
+def select_subdomain(ds: xr.Dataset, bbox: str):
     """Limit the data to a subdomain."""
-    if bbox.west < bbox.east:
+    west, south, east, north = [float(p) for p in bbox.split(", ")]
+    if west < east:
         return ds.sel(
-            longitude=slice(bbox.west, bbox.east),  # -180 to 180
-            latitude=slice(bbox.north, bbox.south),  # 90 - -90 (decreasing)
+            longitude=slice(west, east),  # -180 to 180
+            latitude=slice(north, south),  # 90 - -90 (decreasing)
             drop=True,
         )
     else:
         # Need to roll coordinates to maintain a continuous longitude
         # Roll the first lon that's NOT in the bbox to the front of the array
         # That way, all values that ARE in the bbox, must be contiguous.
-        in_bbox = (ds.longitude < bbox.east) | (ds.longitude > bbox.west)
+        in_bbox = (ds.longitude < east) | (ds.longitude > west)
         n_roll = in_bbox.argmin().item()  # first false value
         ds_rolled = ds.roll(longitude=-n_roll, roll_coords=True)
-        in_bbox = (ds_rolled.longitude < bbox.east) | (ds_rolled.longitude > bbox.west)
-        return ds_rolled.sel(longitude=in_bbox, latitude=slice(bbox.north, bbox.south))
+        in_bbox = (ds_rolled.longitude < east) | (ds_rolled.longitude > west)
+        return ds_rolled.sel(longitude=in_bbox, latitude=slice(north, south))
+
+
+def read_or_retry(filename: str) -> xr.Dataset:
+    """Read a file, retrying if it fails.
+
+    Avoids spurious errors in decoding time when using OpenDAP
+
+    Retry behaviour copied from xarray.backends.common.robust_getitem
+    """
+    max_retries = 8  # ~ 4 minutes maximum, so far always worked within 6 tries.
+    initial_delay = 500  # milliseconds
+
+    for n in range(max_retries + 1):
+        try:
+            return xr.open_dataset(
+                filename, engine="netcdf4", cache=False, use_cftime=True
+            )
+        except IndexError:
+            if n == max_retries:
+                raise
+            base_delay = initial_delay * 2**n
+            next_delay = base_delay + np.random.randint(base_delay)
+            msg = (
+                f"file read failed, waiting {next_delay} ms before trying again "
+                f"({max_retries - n} tries remaining). {traceback.format_exc(limit=2, chain=False)}"
+            )
+            logger.debug(msg)
+            time.sleep(1e-3 * next_delay)
+
+
+@lru_cache(maxsize=6)
+def _load_slice_with_cache(filename, time, subset: str, bbox: Optional[str] = None):
+    """Load data slice from file."""
+    variables = {
+        "fluxes": ["fx_upper", "fx_lower", "fy_upper", "fy_lower", "evap", "precip"],
+        "states": ["s_upper", "s_lower"],
+    }[subset]
+
+    logger.debug(f"Loading {subset} at time {time} from {filename}")
+
+    ds = read_or_retry(filename)
+    ds = ds.sel(time=time).drop("time").squeeze()
+    ds = ds[variables]
+
+    if bbox is not None:
+        ds = select_subdomain(ds, bbox)
+
+    return fully_load_and_detach(ds)
+
+
+def load_slice(t: CfDateTime, subset: str, config: Config) -> xr.Dataset:
+    filename = input_path(t, config.preprocessed_data_folder)
+    time = t.strftime("%Y-%m-%dT%H:%M:%S")
+    domain = str(config.tracking_domain) if config.tracking_domain else None
+
+    # Avoid modifying the cached slice!
+    slice = _load_slice_with_cache(filename, time, subset, domain)
+    return deepcopy(slice)
 
 
 def load_data(t: CfDateTime, config: Config, subset: str = "fluxes") -> xr.Dataset:
     """Load variable at t, interpolate if needed."""
-    variables = {
-        "fluxes": ["fx_upper", "fx_lower", "fy_upper", "fy_lower", "evap", "precip"],
-        "states": ["s_upper", "s_lower"],
-    }
-
-    t1 = round_cftime(t, config.input_frequency, "ceil")
-    da1 = read_data_at_time(t1, config)[variables[subset]]
-    if t == t1:
-        # this saves a lot of work if times are aligned with input
-        return da1
-
     t0 = round_cftime(t, config.input_frequency, "floor")
-    da0 = read_data_at_time(t0, config)[variables[subset]]
-    if t == t0:
-        return da0
+    t1 = round_cftime(t, config.input_frequency, "ceil")
 
+    if t == t0 or t == t1:
+        return load_slice(t, subset, config)
+
+    da0 = load_slice(t0, subset, config)
+    da1 = load_slice(t1, subset, config)
     return da0 + (t - t0) / (t1 - t0) * (da1 - da0)
 
 
@@ -106,7 +148,7 @@ def load_tagging_region(config: Config, t: Optional[CfDateTime] = None):
         tagging_region = xr.open_dataarray(config.tagging_region)
 
     if config.tracking_domain is not None:
-        tagging_region = select_subdomain(tagging_region, config.tracking_domain)
+        tagging_region = select_subdomain(tagging_region, str(config.tracking_domain))
 
     if t is not None:
         return tagging_region.sel(time=t, method="nearest")