+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
 - updated reference to WAM2layers v3 paper
+- support for using pre-processed data via OpenDAP ([#475](https://github.com/WAM2layers/WAM2layers/pull/451))
 - release instructions template
 
 ### Fixed